+++ core/__init__.py
@@ -1,0 +1,1 @@
+"""Modul inti Dat E (desktop + HTTP server)."""